requirements.txt
flask==3.0.0
requests==2.31.0
numpy==1.26.4
matplotlib==3.9.3
pandas==2.2.3
seaborn==0.13.2
wordcloud==1.9.3
reportlab==4.0.7
python-docx==1.1.0
easyocr==1.7.1
Pillow==10.1.0
opencv-python==4.10.0.84
